adding timer and logic in timer component
