+++ client/src/components/QuestionBox.vue
@@ -4,21 +4,18 @@
     v-show="isReady"
     >
       <b-card-text class="">{{questions}}</b-card-text>
-    <circular-count-down-timer
-    :initial-value="10"
-    :steps="7"
-    :size="120"
-    :padding="10"
-    :paused="!isReady"
-    ></circular-count-down-timer>
     </b-card>
-
+    <Timer/>
   </div>
 </template>
 
 <script>
 
+import Timer from './Timer.vue'
 export default {
+  components: {
+    Timer
+  },
   data () {
     return {
       round: 10
@@ -27,15 +24,6 @@ export default {
   computed: {
     isReady () {
       return this.$store.state.isReady
-    },
-    isTimeOut () {
-      return this.$store.state.isReady
-    },
-    questions () {
-      const index = Math.floor(Math.random() * this.$store.state.questions.length)
-      let question = this.$store.state.questions[index]
-      console.log(index, question)
-      return question
     }
   }
 }
