+++ client/src/components/TrueStatement.vue
@@ -1,0 +1,17 @@
+<template>
+  <b-card class="text-center position-sticky w-50 h-auto"
+    v-show="isWinner"
+    >
+      <b-card-text class="">{{questions}}</b-card-text>
+    </b-card>
+</template>
+
+<script>
+export default {
+  
+}
+</script>
+
+<style>
+
+</style>
