+++ client/src/store/index.js
@@ -12,7 +12,9 @@ export default new Vuex.Store({
     // divider
     isReady: false,
     timeOut: false,
+    setTime: 10,
     questions: [],
+    isWinner: false
   },
   mutations: {
     emptyUserId(state) {
